naming refactoring & actions as string into stores
