+++ app/js/fluxo/index.js
@@ -1,8 +1,10 @@
 
-var action = require('./action');
-var store = require('./store');
+var actionsFactory = require('./actions-factory');
 var actionsRegister = require('./actions-register');
 
-exports.createAction = action.create;
-exports.createStore = store.create;
+var storeFactory = require('./store-factory');
+
+exports.createAction = actionsFactory.create;
+exports.createStore = storeFactory.create;
+
 exports.runAction = actionsRegister.run;
